+++ cores/api-spring/src/test/java/com/enistere/core/EnistereCoreApplicationTests.java
@@ -1,12 +1,8 @@
 package com.enistere.core;
 
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
 
-@SpringBootTest
-@ActiveProfiles("test")
-class EnistereCoreApplicationTests {
+class EnistereCoreApplicationTests extends AbstractIntegrationTest {
 
     @Test
     void contextLoads() {